Moving hex assembly-specific tools out of FuelHandler
